Fixing the Jyson library
The library will now handle functions that return undefined.
As well we will make sure that if a value returns false then
we will not exit the loop.
