+++ lib/jyson.js
@@ -17,9 +17,8 @@ const getKey = (object, key) => {
 };
 
 const fillKeys = (json, object, opts) => {
-  const jsonKeys = Object.keys(json);
-
-  _.each(jsonKeys, key => {
+  const jsonResult = {};
+  Object.keys(json).forEach(key => {
     // String Path
     if(_.isString(json[key])){
       const value = getKey(object, json[key], null);
@@ -28,30 +27,31 @@ const fillKeys = (json, object, opts) => {
       assert.ok(!_.isArray(value), `jyson encountered an array when it was not expecting one: ${json[key]}`);
 
       // otherwise return the value
-      return json[key] = value;
+      return jsonResult[key] = value;
     }
 
     // Array
     if(_.isArray(json[key])) {
       assert.ok(json[key].length === 1, `jyson template arrays must be of length one: ${json[key]}`);
 
-      return json[key] = getKey(object, json[key][0], null);
+      return jsonResult[key] = getKey(object, json[key][0], null);
     }
 
     if (_.isFunction(json[key])) {
-      return json[key] = json[key]({object, key, opts});
+      const functionResult = jsonResult[key] = json[key]({object, key, opts});
+      return jsonResult[key] = (typeof functionResult === 'undefined' ? null : functionResult);
     }
 
     // Nested Object
     if(_.isObject(json[key])) {
-      return json[key] = fillKeys(json[key], object, opts);
+      return jsonResult[key] = fillKeys(json[key], object, opts);
     }
 
     // We should not be able to get here
     throw `jyson encountered an unknown template value: ${json[key]}`;
   });
 
-  return json;
+  return jsonResult;
 };
 
 module.exports.buildTemplateFunction = (templateObject) => {
@@ -64,9 +64,7 @@ module.exports.buildTemplateFunction = (templateObject) => {
     return objects.map(object => {
       const json = _.cloneDeep(templateObject);
 
-      fillKeys(json, object, opts);
-
-      return json;
+      return fillKeys(json, object, opts);
     });
 
   };
