+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "jyson",
-  "version": "1.0.5",
+  "version": "1.0.6",
   "description": "A template engine for json.",
   "main": "index.js",
   "scripts": {
