+++ spec/lib/jyson/jyson.function.spec.js
@@ -58,30 +58,30 @@ describe('jyson.function.spec: a function in the template', () => {
   });
 
   describe('simple undefined functions', () => {
-      beforeEach(() =>{
-          this.templateFunction = jyson.buildTemplateFunction({
-              a: () => 1,
-              b: () => {
-                  return;
-              },
-              c: function() {
-                  return;
-              }
-          });
+    beforeEach(() =>{
+      this.templateFunction = jyson.buildTemplateFunction({
+        a: () => 1,
+        b: () => {
+          return;
+        },
+        c: function() {
+          return;
+        }
       });
+    });
 
-      it('must convert an object to "json"', () => {
-          const input = {
-              a: 0,
-              b: 0,
-              c: 0
-          };
-          const json = this.templateFunction(input);
+    it('must convert an object to "json"', () => {
+      const input = {
+        a: 0,
+        b: 0,
+        c: 0
+      };
 
-          expect(json).to.have.all.keys('a');
-          expect(json.a).to.equal(1);
-
-      });
+      const json = this.templateFunction(input);
+      expect(json.a).to.equal(1);
+      expect(json.b).to.equal(null);
+      expect(json.c).to.equal(null);
+    });
   });
 
   describe('nested comments via functions', () => {
